Split ChromiumResolver to 3 files
