+++ src/services/chromium/context.ts
@@ -1,0 +1,4 @@
+type ChromiumContext = {
+  directory: string;
+  executable: string;
+};
